+++ src/utils/storage.js
@@ -10,12 +10,12 @@ function getDefaultDirectory() {
     }
 }
 
-async function load(directory = getDefaultDirectory(), type) {
+async function load(directory = getDefaultDirectory()) {
     if (env.isNode()) {
-        const file = path.join(directory, type + '.json');
+        const file = path.join(directory, 'chlu.json');
         if (!fs.existsSync(file)) return {};
         return await new Promise((resolve, reject) => {
-            fs.readFile(path.join(directory, type + '.json'), (err, result) => {
+            fs.readFile(path.join(directory, 'chlu.json'), (err, result) => {
                 if (err) reject(err);
                 try {
                     const data = JSON.parse(result.toString());
@@ -26,21 +26,21 @@ async function load(directory = getDefaultDirectory(), type) {
             });
         });
     } else {
-        const string = localStorage.getItem('chlu-' + type + '-data') || '{}';
+        const string = localStorage.getItem('chlu-data') || '{}';
         return JSON.parse(string);
     }
 }
 
-async function save(directory = getDefaultDirectory(), data, type) {
+async function save(directory = getDefaultDirectory(), data) {
     const string = JSON.stringify(data);
     if (env.isNode()) {
         if (!fs.existsSync(directory)) fs.mkdirSync(directory);
-        const file = path.join(directory, type + '.json');
+        const file = path.join(directory, 'chlu.json');
         await new Promise((resolve, reject) => {
             fs.writeFile(file, string, err => err ? reject(err) : resolve());
         });
     } else {
-        localStorage.setItem('chlu-' + type + '-data', string);
+        localStorage.setItem('chlu-data', string);
     }
 }