+++ tests/Crypto.test.js
@@ -11,7 +11,6 @@ describe('Crypto Module', () => {
 
     beforeEach(async () => {
         chluIpfs = new ChluIPFS({
-            type: ChluIPFS.types.vendor,
             enablePersistence: false,
             cache: { enabled: false },
             logger: logger('Vendor')